summit-pine-ui/requirements.txt
streamlit==1.43.2
psycopg2-binary==2.9.10
pandas==2.2.3
plotly==5.24.1
pypdf>=4.0.0
requests>=2.31.0
